pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "TEMPLATE_PROJECT"
version = "0.0.0"
description = "Python project template scaffold extracted from bristol_ml — Hydra+Pydantic config, Claude Code agent infrastructure, four-tier doc methodology."
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [{ name = "Sam Drew", email = "comm@sdrew.com" }]
dependencies = [
    # Config: Hydra composes YAML; Pydantic validates the resolved tree
    # at the CLI boundary.  See ADR 0001.
    "hydra-core>=1.3,<2",
    "omegaconf>=2.3,<3",
    "pydantic>=2.7,<3",
    # Logging.
    "loguru>=0.7,<1",
]

[dependency-groups]
dev = [
    "pytest>=8,<9",
    "ruff>=0.6,<1",
    "pre-commit>=3.7,<5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/TEMPLATE_PROJECT"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["TEMPLATE_PROJECT", "conf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
addopts = "-ra --strict-markers"
markers = [
    "slow: long-running tests excluded from the default run; opt in with -m slow.",
]
